feat: Update list item functionality in List component
The List component has been updated to include the ability to add, edit, and delete list items. This was achieved by adding new state variables and implementing corresponding event handlers. The changes also include updates to the API endpoints used for fetching and updating list data.

Note: This commit message follows the established convention of using a prefix to indicate the type of change (feat for feature).